requirements.txt
gradio
numpy==1.23.5
praat-parselmouth>=0.4.2
pyworld>=0.3.2
faiss-cpu==1.7.3
torchcrepe==0.0.20
ffmpeg-python>=0.2.0
fairseq==0.12.2
scipy==1.9.3
librosa==0.9.1
yt_dlp
demucs